pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "affine-gaps"
version = "0.1.0"
description = "Numba-accelerated Python implementation of affine gap penalty extensions for Needleman-Wunsch and Smith-Waterman algorithms"
readme = { file = "README.md", content-type = "text/markdown" }
authors = [
    { name = "Ash Vardanian", email = "1983160+ashvardanian@users.noreply.github.com" },
]
requires-python = ">=3.6"
dependencies = ["numba", "numpy", "colorama"]
keywords = ["bioinformatics", "sequence alignment", "numba", "affine gaps"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
]

[project.optional-dependencies]
dev = ["biopython", "stringzilla", "pytest", "pytest-repeat"]

[tool.hatch.entry-points.scripts]
affine-gaps = "affine_gaps:main"
